made create_activity handle
