+++ core/internal/application/handlers/get_all_sport_section.go
@@ -1,4 +1,4 @@
-package sports
+package handlers
 
 import (
 	"context"
