+++ core/internal/application/handlers/get_teams_by_activity.go
@@ -1,8 +1,7 @@
-package teams
+package handlers
 
 import (
 	"context"
-	"match/internal/application/handlers"
 	"match/internal/application/handlers/mappers"
 	"match/internal/domain/dto"
 	"match/internal/generated/server"
@@ -36,7 +35,7 @@ func (h *GetTeamsByActivityIDHandler) GetTeamsByActivityID(ectx echo.Context, ac
 	teams, err := h.teamService.GetTeamsByActivityID(ctx, activityID)
 	if err != nil {
 		infra.Errorf(ctx, "Internal server error while trying to find activity: %v", err)
-		return handlers.InternalErrorResponse(ectx, err.Error())
+		return InternalErrorResponse(ectx, err.Error())
 	}
 	infra.Infof(ctx, "%d teams have been found and extracted succesfully", 1)
